bt ret obj ptr

+++ tt1/ex_cl.hpp
@@ -4,30 +4,30 @@
 
 class ExComp{
     private:
-        int             val;
+        double          val;
         unsigned int    ID;
         std::string     txt;
     public:
         static int      lastID;
         static int      duplicates;
         ExComp(void){};
-        ExComp(int v, const std::string& input);
+        ExComp(double v, const std::string& input);
         ~ExComp(){};
         void    print_txt();
-        int     ret_val()  {return val;};
+        double  ret_val()  {return val;};
         bool    operator== (const ExComp& cmp);
-        bool    operator== (const int cmp);
+        bool    operator== (const double cmp);
         bool    operator<  (const ExComp& cmp);
-        bool    operator<  (const int cmp);
+        bool    operator<  (const double cmp);
         bool    operator>  (const ExComp& cmp);
-        bool    operator>  (const int cmp);
+        bool    operator>  (const double cmp);
         friend  std::ostream& operator<<(std::ostream& os, const ExComp& prnt);
 };
 
 int ExComp::lastID = 0;
 int ExComp::duplicates = 0;
 
-ExComp::ExComp(int v, const std::string& input){
+ExComp::ExComp(double v, const std::string& input){
     this->val = v;
     this->txt = input;
     lastID++;
@@ -37,7 +37,7 @@ ExComp::ExComp(int v, const std::string& input){
 void ExComp::print_txt(){
     std::cout<< "Obj.ID=" << this->ID << "|Text=" << this->txt << std::endl; }
 
-bool ExComp::operator==(const int cmp){
+bool ExComp::operator==(const double cmp){
     std::cout<< this->val << " vs " << cmp << std::endl;
     if (this->val == cmp) duplicates++;
     return this->val == cmp; }
@@ -50,13 +50,13 @@ bool ExComp::operator==(const ExComp& cmp){
 bool ExComp::operator< (const ExComp& cmp){
     return this->val < cmp.val; }
 
-bool ExComp::operator< (const int cmp){
+bool ExComp::operator< (const double cmp){
     return this->val < cmp; }
 
 bool ExComp::operator> (const ExComp& cmp){
     return this->val > cmp.val; }
 
-bool ExComp::operator> (const int cmp){
+bool ExComp::operator> (const double cmp){
     return this->val > cmp; }
 
 std::ostream& operator<<(std::ostream& os, const ExComp& prnt){
